Split resolver metadata collection helpers
